// client/src/store/trending/constants.ts
export enum TrendingTypes {
  REQUEST_TRENDING_MOVIES = 'REQUEST_TRENDING',
  FETCH_TRENDING_MOVIES_START = 'FETCH_TRENDING_MOVIES_START',
  FETCH_TRENDING_MOVIES_SUCCESS = 'FETCH_TRENDING_MOVIES_SUCCESS',
  FETCH_TRENDING_MOVIES_ERROR = 'FETCH_TRENDING_MOVIES_ERROR',

  REQUEST_TRENDING_SHOWS = 'REQUEST_TRENDING',
  FETCH_TRENDING_SHOWS_START = 'FETCH_TRENDING_SHOWS_START',
  FETCH_TRENDING_SHOWS_SUCCESS = 'FETCH_TRENDING_SHOWS_SUCCESS',
  FETCH_TRENDING_SHOWS_ERROR = 'FETCH_TRENDING_SHOWS_ERROR',
}